+++ src/lib.rs
@@ -112,10 +112,40 @@ pub struct OccupiedEntry<'a, Value> {
 }
 
 /// An iterator over an `LruCache`'s entries that updates the timestamps as values are traversed.
+/// Values are produced in the most recently used order.
 pub struct Iter<'a, Key: 'a, Value: 'a> {
-    map_iter_mut: btree_map::IterMut<'a, Key, (Value, Instant)>,
+    /// Reference to the iterated cache.
+    map: &'a mut BTreeMap<Key, (Value, Instant)>,
+    /// Ordered cache entry keys where the least recently used items are first.
     list: &'a mut VecDeque<Key>,
     lru_cache_ttl: Option<Duration>,
+    /// Index in `list` of the previously used item.
+    item_index: usize,
+}
+
+impl<'a, Key, Value> Iter<'a, Key, Value>
+where
+    Key: Ord,
+{
+    /// Returns next unexpired item in the cache or `None` if no such items.
+    /// Expired items are removed from the cache.
+    fn next_unexpired(&mut self, now: Instant) -> Option<Key> {
+        loop {
+            self.item_index = self.item_index.checked_sub(1)?;
+            let key = self.list.remove(self.item_index)?;
+            let value = self.map.get(&key)?;
+
+            if let Some(ttl) = self.lru_cache_ttl {
+                if value.1 + ttl > now {
+                    return Some(key);
+                } else {
+                    let _ = self.map.remove(&key);
+                }
+            } else {
+                return Some(key);
+            }
+        }
+    }
 }
 
 impl<'a, Key, Value> Iterator for Iter<'a, Key, Value>
@@ -124,20 +154,74 @@ where
 {
     type Item = (&'a Key, &'a Value);
 
+    /// Returns the next element in the cache and moves it to the top of the cache.
+    /// The most recently used items are yield first.
+    #[allow(unsafe_code)]
     fn next(&mut self) -> Option<(&'a Key, &'a Value)> {
         let now = Instant::now();
-        let not_expired = match self.lru_cache_ttl {
-            Some(ttl) => self
-                .map_iter_mut
-                .find(|&(_, &mut (_, instant))| instant + ttl > now),
-            None => self.map_iter_mut.next(),
-        };
+        let key = self.next_unexpired(now)?;
+        self.list.push_back(key);
+        let key = self.list.back()?;
+        let mut value = self.map.get_mut(&key)?;
+        value.1 = now;
 
-        not_expired.map(|(key, &mut (ref value, ref mut instant))| {
-            LruCache::<Key, Value>::update_key(self.list, key);
-            *instant = now;
-            (key, value)
-        })
+        unsafe {
+            let key = std::mem::transmute::<&Key, &'a Key>(key);
+            let value = std::mem::transmute::<&Value, &'a Value>(&value.0);
+            Some((key, value))
+        }
+    }
+}
+
+/// Entry produced by `NotifyIter` that might be still valid or expired.
+pub enum TimedEntry<'a, Key: 'a, Value: 'a> {
+    /// Entry has not yet expired.
+    Valid(&'a Key, &'a Value),
+    /// Entry got expired and was evicted from the cache.
+    Expired(Key, Value),
+}
+
+/// Much like `Iter` except will produce expired entries too where `Iter` silently drops them.
+pub struct NotifyIter<'a, Key: 'a, Value: 'a> {
+    /// Reference to the iterated cache.
+    map: &'a mut BTreeMap<Key, (Value, Instant)>,
+    /// Ordered cache entry keys where the least recently used items are first.
+    list: &'a mut VecDeque<Key>,
+    lru_cache_ttl: Option<Duration>,
+    /// Index in `list` of the previously used item.
+    item_index: usize,
+}
+
+impl<'a, Key, Value> Iterator for NotifyIter<'a, Key, Value>
+where
+    Key: Ord + Clone,
+{
+    type Item = TimedEntry<'a, Key, Value>;
+
+    /// Returns the next element in the cache and moves it to the top of the cache.
+    /// The most recently used items are yield first.
+    #[allow(unsafe_code)]
+    fn next(&mut self) -> Option<Self::Item> {
+        self.item_index = self.item_index.checked_sub(1)?;
+        let key = self.list.remove(self.item_index)?;
+        let mut value = self.map.get_mut(&key)?;
+        let now = Instant::now();
+
+        if let Some(ttl) = self.lru_cache_ttl {
+            if value.1 + ttl <= now {
+                let value = self.map.remove(&key)?;
+                return Some(TimedEntry::Expired(key, value.0));
+            }
+        }
+
+        self.list.push_back(key);
+        let key = self.list.back()?;
+        value.1 = now;
+        unsafe {
+            let key = std::mem::transmute::<&Key, &'a Key>(key);
+            let value = std::mem::transmute::<&Value, &'a Value>(&value.0);
+            Some(TimedEntry::Valid(key, value))
+        }
     }
 }
 
@@ -214,22 +298,31 @@ where
     ///
     /// If the key already existed in the cache, the existing value is returned and overwritten in
     /// the cache.  Otherwise, the key-value pair is inserted and `None` is returned.
-    pub fn insert(&mut self, key: Key, value: Value) -> Option<Value> {
+    /// Evicts and returns expired entries.
+    pub fn notify_insert(&mut self, key: Key, value: Value) -> (Option<Value>, Vec<(Key, Value)>) {
+        let expired = self.remove_expired();
+
         if self.map.contains_key(&key) {
             Self::update_key(&mut self.list, &key);
         } else {
-            self.remove_expired();
-            if self.map.len() >= self.capacity {
-                for key in self.list.drain(..=self.map.len() - self.capacity) {
-                    assert!(self.map.remove(&key).is_some());
-                }
-            }
+            self.remove_lru();
             self.list.push_back(key.clone());
-        }
+        };
 
-        self.map
-            .insert(key, (value, Instant::now()))
-            .map(|pair| pair.0)
+        (
+            self.map
+                .insert(key, (value, Instant::now()))
+                .map(|pair| pair.0),
+            expired,
+        )
+    }
+
+    /// Inserts a key-value pair into the cache.
+    ///
+    /// If the key already existed in the cache, the existing value is returned and overwritten in
+    /// the cache.  Otherwise, the key-value pair is inserted and `None` is returned.
+    pub fn insert(&mut self, key: Key, value: Value) -> Option<Value> {
+        self.notify_insert(key, value).0
     }
 
     /// Removes a key-value pair from the cache.
@@ -252,6 +345,16 @@ where
     pub fn clear(&mut self) {
         self.map.clear();
         self.list.clear();
+    }
+
+    /// Much like `get()`, except in addition returns expired entries.
+    pub fn notify_get<Q: ?Sized>(&mut self, key: &Q) -> (Option<&Value>, Vec<(Key, Value)>)
+    where
+        Key: Borrow<Q>,
+        Q: Ord,
+    {
+        let (value, expired) = self.notify_get_mut(key);
+        (value.map(|v| &*v), expired)
     }
 
     /// Retrieves a reference to the value stored under `key`, or `None` if the key doesn't exist.
@@ -282,20 +385,33 @@ where
     }
 
     /// Retrieves a mutable reference to the value stored under `key`, or `None` if the key doesn't
+    /// exist. Also removes expired elements and updates the time.
+    pub fn notify_get_mut<Q: ?Sized>(&mut self, key: &Q) -> (Option<&mut Value>, Vec<(Key, Value)>)
+    where
+        Key: Borrow<Q>,
+        Q: Ord,
+    {
+        let expired = self.remove_expired();
+
+        let list = &mut self.list;
+        (
+            self.map.get_mut(key).map(|result| {
+                Self::update_key(list, key);
+                result.1 = Instant::now();
+                &mut result.0
+            }),
+            expired,
+        )
+    }
+
+    /// Retrieves a mutable reference to the value stored under `key`, or `None` if the key doesn't
     /// exist.  Also removes expired elements and updates the time.
     pub fn get_mut<Q: ?Sized>(&mut self, key: &Q) -> Option<&mut Value>
     where
         Key: Borrow<Q>,
         Q: Ord,
     {
-        self.remove_expired();
-
-        let list = &mut self.list;
-        self.map.get_mut(key).map(|result| {
-            Self::update_key(list, key);
-            result.1 = Instant::now();
-            &mut result.0
-        })
+        self.notify_get_mut(key).0
     }
 
     /// Returns whether `key` exists in the cache or not.
@@ -351,11 +467,27 @@ where
 
     /// Returns an iterator over all entries that updates the timestamps as values are
     /// traversed. Also removes expired elements before creating the iterator.
-    pub fn iter(&mut self) -> Iter<'_, Key, Value> {
-        self.remove_expired();
+    /// Values are produced in the most recently used order.
+    ///
+    /// Also, evicts and returns expired entries.
+    pub fn notify_iter(&mut self) -> NotifyIter<Key, Value> {
+        NotifyIter {
+            item_index: self.list.len(),
+            map: &mut self.map,
+            list: &mut self.list,
+            lru_cache_ttl: self.time_to_live,
+        }
+    }
+
+    /// Returns an iterator over all entries that updates the timestamps as values are
+    /// traversed. Also removes expired elements before creating the iterator.
+    /// Values are produced in the most recently used order.
+    pub fn iter(&mut self) -> Iter<Key, Value> {
+        let _ = self.remove_expired();
 
         Iter {
-            map_iter_mut: self.map.iter_mut(),
+            item_index: self.list.len(),
+            map: &mut self.map,
             list: &mut self.list,
             lru_cache_ttl: self.time_to_live,
         }
@@ -380,19 +512,38 @@ where
         }
     }
 
-    fn remove_expired(&mut self) {
+    /// If expiry timeout is set, removes expired items from the cache and returns them.
+    fn remove_expired(&mut self) -> Vec<(Key, Value)> {
         let (map, list) = (&mut self.map, &mut self.list);
-        if let Some((i, val)) = self.time_to_live.and_then(|ttl| {
-            list.iter()
-                .enumerate()
-                .filter_map(|(i, key)| map.remove(key).map(|val| (i, val)))
-                .find(|&(_, (_, t))| t + ttl >= Instant::now())
-        }) {
-            // we have found one item not expired, we must insert it back
-            let _ = map.insert(list[i].clone(), val);
-            let _ = list.drain(..i);
+
+        if let Some(ttl) = self.time_to_live {
+            let mut expired_values = Vec::new();
+            for key in list.iter() {
+                if map[key].1 + ttl >= Instant::now() {
+                    break;
+                }
+                if let Some(entry) = map.remove(key) {
+                    expired_values.push(entry.0);
+                }
+            }
+            // remove keys as well
+            return list
+                .drain(..expired_values.len())
+                .zip(expired_values)
+                .collect();
         } else if map.is_empty() {
             list.clear();
+        }
+
+        Vec::new()
+    }
+
+    /// Removes least recently used items to make space for new ones.
+    fn remove_lru(&mut self) {
+        if self.map.len() >= self.capacity {
+            for key in self.list.drain(..=self.map.len() - self.capacity) {
+                assert!(self.map.remove(&key).is_some());
+            }
         }
     }
 }
@@ -449,8 +600,8 @@ impl<'a, Key: Ord + Clone, Value> Entry<'a, Key, Value> {
 
 #[cfg(test)]
 mod test {
-    use rand::distributions::{Distribution, Standard};
-    use rand::thread_rng;
+    use super::*;
+    use rand;
     use std::time::Duration;
 
     #[cfg(feature = "fake_clock")]
@@ -610,34 +761,193 @@ mod test {
         assert_eq!(lru_cache.len(), 1);
     }
 
-    #[test]
-    fn iter() {
-        let mut lru_cache = super::LruCache::<usize, usize>::with_capacity(3);
+    mod notify_insert {
+        use super::*;
 
-        let _ = lru_cache.insert(0, 0);
-        sleep(1);
-        let _ = lru_cache.insert(1, 1);
-        sleep(1);
-        let _ = lru_cache.insert(2, 2);
-        sleep(1);
+        #[test]
+        fn it_removes_expired_entries() {
+            let ttl = Duration::from_millis(200);
+            let mut lru_cache = LruCache::<usize, usize>::with_expiry_duration(ttl);
+            let _ = lru_cache.insert(1, 1);
+            let _ = lru_cache.insert(2, 2);
+            sleep(250);
 
-        assert_eq!(
-            vec![(&0, &0), (&1, &1), (&2, &2)],
-            lru_cache.iter().collect::<Vec<_>>()
-        );
+            let _ = lru_cache.notify_insert(3, 3);
 
-        let initial_instant0 = lru_cache.map[&0].1;
-        let initial_instant2 = lru_cache.map[&2].1;
-        sleep(1);
+            assert_eq!(lru_cache.map.len(), 1);
+            assert_eq!(lru_cache.map[&3].0, 3);
+        }
 
-        // only the first two entries should have their timestamp updated (and position in list)
-        let _ = lru_cache.iter().take(2).all(|_| true);
+        #[test]
+        fn it_returns_removed_expired_entries() {
+            let ttl = Duration::from_millis(200);
+            let mut lru_cache = LruCache::<usize, usize>::with_expiry_duration(ttl);
+            let _ = lru_cache.insert(1, 1);
+            let _ = lru_cache.insert(2, 2);
+            sleep(250);
 
-        assert_ne!(lru_cache.map[&0].1, initial_instant0);
-        assert_eq!(lru_cache.map[&2].1, initial_instant2);
+            let (_replaced, expired) = lru_cache.notify_insert(3, 3);
 
-        assert_eq!(*lru_cache.list.front().unwrap(), 2);
-        assert_eq!(*lru_cache.list.back().unwrap(), 1);
+            assert_eq!(expired.len(), 2);
+            assert_eq!(expired[0], (1, 1));
+            assert_eq!(expired[1], (2, 2));
+        }
+    }
+
+    mod iter {
+        use super::*;
+
+        #[test]
+        fn it_returns_none_when_cache_is_empty() {
+            let mut lru_cache = LruCache::<usize, usize>::with_capacity(3);
+
+            let next = lru_cache.iter().next();
+
+            assert!(next.is_none());
+        }
+
+        #[test]
+        fn it_updates_item_timestamps_of_traversed_items() {
+            let mut lru_cache = LruCache::<usize, usize>::with_capacity(3);
+            let _ = lru_cache.insert(0, 0);
+            sleep(1);
+            let _ = lru_cache.insert(1, 1);
+            sleep(1);
+            let _ = lru_cache.insert(2, 2);
+            sleep(1);
+
+            let initial_instant0 = lru_cache.map[&0].1;
+            let initial_instant2 = lru_cache.map[&2].1;
+            sleep(1);
+
+            // only the first two entries should have their timestamp updated (and position in list)
+            let _ = lru_cache.iter().take(2).all(|_| true);
+
+            assert_ne!(lru_cache.map[&2].1, initial_instant2);
+            assert_eq!(lru_cache.map[&0].1, initial_instant0);
+        }
+
+        #[test]
+        fn it_moves_traversed_items_to_the_top_of_the_cache() {
+            let mut lru_cache = LruCache::<usize, usize>::with_capacity(3);
+            let _ = lru_cache.insert(0, 0);
+            let _ = lru_cache.insert(1, 1);
+            let _ = lru_cache.insert(2, 2);
+
+            let _ = lru_cache.iter().take(2).all(|_| true);
+
+            assert_eq!(*lru_cache.list.front().unwrap(), 0);
+            assert_eq!(*lru_cache.list.back().unwrap(), 1);
+        }
+
+        #[test]
+        fn it_yields_the_most_recent_items_first() {
+            let mut lru_cache = LruCache::<usize, usize>::with_capacity(4);
+            let _ = lru_cache.insert(2, 2);
+            let _ = lru_cache.insert(0, 0);
+            let _ = lru_cache.insert(3, 3);
+            let _ = lru_cache.insert(1, 1);
+
+            let cached = lru_cache.iter().collect::<Vec<_>>();
+
+            assert_eq!(cached, vec![(&1, &1), (&3, &3), (&0, &0), (&2, &2)]);
+        }
+
+        #[test]
+        fn it_removes_expired_items() {
+            let mut lru_cache =
+                LruCache::<usize, usize>::with_expiry_duration(Duration::from_millis(3));
+            let _ = lru_cache.insert(0, 0);
+            sleep(1);
+            let _ = lru_cache.insert(1, 1);
+            sleep(4);
+            let _ = lru_cache.insert(2, 2);
+
+            let items: Vec<_> = lru_cache.iter().collect();
+
+            assert_eq!(items.len(), 1);
+            assert_eq!(items[0], (&2, &2));
+        }
+    }
+
+    mod notify_iter {
+        use super::*;
+
+        #[test]
+        fn it_returns_none_when_cache_is_empty() {
+            let mut lru_cache = LruCache::<usize, usize>::with_capacity(3);
+
+            let next = lru_cache.notify_iter().next();
+
+            assert!(next.is_none());
+        }
+
+        #[test]
+        fn it_yields_the_most_recent_items_first() {
+            let mut lru_cache = LruCache::<usize, usize>::with_capacity(4);
+            let _ = lru_cache.insert(2, 2);
+            let _ = lru_cache.insert(0, 0);
+            let _ = lru_cache.insert(3, 3);
+            let _ = lru_cache.insert(1, 1);
+
+            let cached = lru_cache
+                .notify_iter()
+                .map(|entry| match entry {
+                    TimedEntry::Valid(key, value) => (key, value),
+                    _ => panic!("Unexpected expired entry"),
+                })
+                .collect::<Vec<_>>();
+
+            assert_eq!(cached, vec![(&1, &1), (&3, &3), (&0, &0), (&2, &2)]);
+        }
+
+        #[test]
+        fn it_produces_expired_and_valid_entries() {
+            let mut lru_cache =
+                LruCache::<usize, usize>::with_expiry_duration(Duration::from_millis(300));
+            let _ = lru_cache.insert(0, 0);
+            let _ = lru_cache.insert(1, 1);
+            sleep(250);
+            let _ = lru_cache.insert(2, 2);
+            sleep(60);
+
+            let expired: Vec<_> = lru_cache
+                .notify_iter()
+                .filter_map(|entry| match entry {
+                    TimedEntry::Expired(key, value) => Some((key, value)),
+                    _ => None,
+                })
+                .collect();
+            let valid: Vec<_> = lru_cache
+                .notify_iter()
+                .filter_map(|entry| match entry {
+                    TimedEntry::Valid(key, value) => Some((key, value)),
+                    _ => None,
+                })
+                .collect();
+
+            assert_eq!(valid.len(), 1);
+            assert_eq!(valid[0], (&2, &2));
+            assert_eq!(expired.len(), 2);
+            assert_eq!(expired[0], (1, 1));
+            assert_eq!(expired[1], (0, 0));
+        }
+
+        #[test]
+        fn it_removes_expired_items() {
+            let mut lru_cache =
+                LruCache::<usize, usize>::with_expiry_duration(Duration::from_millis(300));
+            let _ = lru_cache.insert(0, 0);
+            let _ = lru_cache.insert(1, 1);
+            sleep(250);
+            let _ = lru_cache.insert(2, 2);
+            sleep(60);
+
+            let _items: Vec<_> = lru_cache.notify_iter().collect();
+
+            assert!(lru_cache.get(&0).is_none());
+            assert!(lru_cache.get(&1).is_none());
+        }
     }
 
     #[test]
@@ -694,5 +1004,62 @@ mod test {
         assert_eq!(Some(&mut 0), lru_cache.get_mut("foo"));
         assert_eq!(Some(&0), lru_cache.peek("foo"));
         assert_eq!(Some(0), lru_cache.remove("foo"));
+    }
+
+    mod remove_expired {
+        use super::*;
+
+        #[test]
+        fn it_removes_expired_entries_from_the_map() {
+            let ttl = Duration::from_millis(200);
+            let mut lru_cache = LruCache::<usize, usize>::with_expiry_duration(ttl);
+            let _ = lru_cache.insert(1, 1);
+            let _ = lru_cache.insert(2, 2);
+            sleep(150);
+            let _ = lru_cache.insert(3, 3);
+            let _ = lru_cache.insert(4, 4);
+            sleep(60);
+
+            let _ = lru_cache.remove_expired();
+
+            assert_eq!(lru_cache.map.len(), 2);
+            assert_eq!(lru_cache.map[&3].0, 3);
+            assert_eq!(lru_cache.map[&4].0, 4);
+        }
+
+        #[test]
+        fn it_removes_expired_entries_from_the_list() {
+            let ttl = Duration::from_millis(200);
+            let mut lru_cache = LruCache::<usize, usize>::with_expiry_duration(ttl);
+            let _ = lru_cache.insert(1, 1);
+            let _ = lru_cache.insert(2, 2);
+            sleep(150);
+            let _ = lru_cache.insert(3, 3);
+            let _ = lru_cache.insert(4, 4);
+            sleep(60);
+
+            let _ = lru_cache.remove_expired();
+
+            assert_eq!(lru_cache.list.len(), 2);
+            assert_eq!(lru_cache.list[0], 3);
+            assert_eq!(lru_cache.list[1], 4);
+        }
+
+        #[test]
+        fn it_returns_expired_entries() {
+            let ttl = Duration::from_millis(200);
+            let mut lru_cache = LruCache::<usize, usize>::with_expiry_duration(ttl);
+            let _ = lru_cache.insert(1, 1);
+            let _ = lru_cache.insert(2, 2);
+            sleep(150);
+            let _ = lru_cache.insert(3, 3);
+            sleep(60);
+
+            let expired = lru_cache.remove_expired();
+
+            assert_eq!(expired.len(), 2);
+            assert_eq!(expired[0], (1, 1));
+            assert_eq!(expired[1], (2, 2));
+        }
     }
 }